LayoutTests/http/tests/dom/iframe-dispatch-load-once-with-document-write.html

<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
promise_test(() => {
    return new Promise((resolve, reject) => {
        var counter = 0;
        var iframe = document.createElement('iframe');
        iframe.src = 'resources/dummy.html';
        iframe.onload = function (ev) {
            try {
                counter++;
                assert_equals(counter, 1);
                iframe.contentDocument.write(`<script ` +  `src="not-found.js">` + `</` + `script>`);
                iframe.contentDocument.close();
            } catch (error) { reject(error); }
            setTimeout(resolve, 100);
        };
        document.body.appendChild(iframe);
    });
}, `iframe should not receive 'load' event more than once`);
</script>
</body>
